Rental Details, At least one field, Add Customer to Inventory Commit

+++ database/migrations/2017_02_021_103649_create_address_table.php
@@ -13,7 +13,7 @@ class CreateAddressTable extends Migration
     {
         Schema::create('address', function (Blueprint $table) {
 
-                        $table->smallInteger('address_id');
+                        $table->increments('address_id');
             $table->string('address', 50);
             $table->string('address2', 50);
             $table->string('district', 20);
