+++ database/migrations/2017_02_021_103649_create_city_table.php
@@ -13,7 +13,7 @@ class CreateCityTable extends Migration
     {
         Schema::create('city', function (Blueprint $table) {
 
-                        $table->smallInteger('city_id');
+                        $table->increments('city_id');
             $table->string('city', 50);
             $table->smallInteger('country_id');
             $table->timestamp('last_update');
